# REST APIs Design/openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Title",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
  "paths": {
...
    "/user/login": {
      "post": {
        "operationId": "login  ",
        "requestBody": {
          "description": "The request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "when user logged in successfully"
          }
        }
      }
    },
...
    "/reviews": {
      "post": {
        "operationId": "submit new review",
        "requestBody": {
          "description": "submit a new review",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/submitReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "the response of new review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/submitReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/:id": {
      "delete": {
        "operationId": "delete review",
        "description": "delete review by its ID",
        "responses": {
          "200": {
            "description": "when the review is deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "result": "True"
                }
              }
            }
          },
          "404": {
            "description": "when the review not found",
            "content": {
              "application/json": {
                "example": {
                  "result": "False"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update review",
        "description": "update review by its Id",
        "requestBody": {
          "description": "the updated data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/submitReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "when the film is undated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/submitReviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "submitReviewRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "filmId": { "type": "integer" },
          "completed": { "type": "boolean", "default": false },
          "reviewDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 10 },
          "review": { "type": "string", "maxLength": 1000 },
          "issuedById": { "type": "integer" }
        }
      },
      "submitReviewResponse": {
        "type": "object",
        "properties": {
          "filmId": { "type": "integer" },
          "issuedById": { "type": "integer" },
          "completed": { "type": "boolean", "default": false },
          "reviewDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 10 },
          "review": { "type": "string", "maxLength": 1000 },
          "id": { "type": "integer" },
          "userId": { "type": "integer" }
        }
      },
...
      "loginResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "email": { "type": "string", "pattern": "email" }
        }
      },
      "loginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "pattern": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"]
      },
      "registerRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "passwordConfirmation": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["email", "password", "passwordConfirmation", "name"]
      },
      "_date": {
        "type": "string",
        "pattern": "^(19|20)[0-9]{2}[- /.](0[1-9]|1[012])[- /.](0[1-9]|[12][0-9]|3[01])$"
      },
...
    }
  }
}